Added a do_configure_append rule to delete .../git/.pc.
Added a do_configure_append rule to delete .../git/.pc because the inclusion of
that directory causes XSCT to fail to create the FSBL libraries on some machines.
We suspect this is due to differences in the order of search/find results. For
example, 'find . -iname lib' on the failing machinev has git/.pc before git/lib,
whereas on a machine where the do_configure rule works, the same 'find' command
shows git/lib before git/.pc. Xilinx SR# 10392136 has been opened to correct this
issue in the Xilinx SDK.