Adjust Linux debug file deploy task for KERNEL_IMAGE_BASE_NAME changes
KERNEL_IMAGE_BASE_NAME used to default to "Image-*", but in morty it defaults
to "-".  Due to this change the debug file deploy tasks were creating a file
with a leading '-' which is undesireable and caused subsequent link operations
to fail.

Fix this by prepending 'vmlinux-' to the filenames derived from this variable.