Bring up Xen bridge after Xen commons
In Krogoth, /proc/xen was being mounted through the fstab and could be accessed
early in the boot process.  This was especially important for the xen networking
bridge which attempts to come up as part of the networking script which is one
of the first init scripts to run.

With the mounting removed from the fstab, the job of mounting this node fell to
xencommons which isn't started until fairly late in the boot process.  This
broke the xen bridge portion of the networking scripts.

To address this we attempt to bring up the xen bridge after xencommons with
a new xenbridge initscript.