ZCU102/U-Boot: read MAC address from eeprom
U-Boot is being built using a common device tree, but this device tree
unfortunately has an incorrect compatible property for the eeprom in the 2019.1
release. This was preventing U-Boot from being able to read the MAC address.

The device tree was fixed upstream, but the fix didn't make it into the Xilinx
2019.1 release. Include the fix here as a patch.

This patch likely won't be necessary here in future releases, since future
releases will likely already include the patch.